fixed cpuid to use uint64 instead of uint32

+++ src/cpuid.c
@@ -10,11 +10,11 @@
 #include <stdint.h>
 #include <stdbool.h>
 // Two defines below from Barry Rountree
-#define MASK_RANGE(m,n) ((((uint32_t)1<<((m)-(n)+1))-1)<<(n))
-#define MASK_VAL(x,m,n) (((uint32_t)(x)&MASK_RANGE((m),(n)))>>(n))
+#define MASK_RANGE(m,n) ((((uint64_t)1<<((m)-(n)+1))-1)<<(n))
+#define MASK_VAL(x,m,n) (((uint64_t)(x)&MASK_RANGE((m),(n)))>>(n))
 
 
-void cpuid(uint32_t leaf, uint32_t *rax, uint32_t *rbx, uint32_t *rcx, uint32_t *rdx)
+void cpuid(uint64_t leaf, uint64_t *rax, uint64_t *rbx, uint64_t *rcx, uint64_t *rdx)
 {
 	asm volatile(
 		"\txchg %%rbx, %%edi\n"
@@ -25,7 +25,7 @@ void cpuid(uint32_t leaf, uint32_t *rax, uint32_t *rbx, uint32_t *rcx, uint32_t 
 		);
 }
 
-void cpuidInput_rax_rcx(uint32_t leafa, uint32_t leafc, uint32_t *rax, uint32_t *rbx, uint32_t *rcx, uint32_t *rdx)
+void cpuidInput_rax_rcx(uint64_t leafa, uint64_t leafc, uint64_t *rax, uint64_t *rbx, uint64_t *rcx, uint64_t *rdx)
 {
 	asm volatile(
 			"xchg %%rbx, %%edi\n"
@@ -48,7 +48,7 @@ void cpuidInput_rax_rcx(uint32_t leafa, uint32_t leafc, uint32_t *rax, uint32_t 
 
 bool cpuid_MPERF_and_APERF()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rcx, 0, 0) == 1) {
@@ -61,7 +61,7 @@ bool cpuid_MPERF_and_APERF()
 
 bool cpuid_timeStampCounter_avail()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 1;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rdx,4,4) == 1) {
@@ -82,7 +82,7 @@ bool cpuid_timeStampCounter_avail()
 
 int cpuid_PMC_num()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 10; // 0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rax, 15, 8);		// This value tells which PMC's are available
@@ -96,7 +96,7 @@ int cpuid_PMC_num()
 
 int cpuid_PERFEVTSEL_num()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 10; // 0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rax, 15,8);		// This value tells which PERFEVTSEL's are available
@@ -111,7 +111,7 @@ int cpuid_PERFEVTSEL_num()
 
 bool cpuid_perf_global_ctrl_EN_PMC()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 10; // 0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 7,0) > 0) {
@@ -124,7 +124,7 @@ bool cpuid_perf_global_ctrl_EN_PMC()
 
 bool cpuid_perf_global_ctrl_EN_FIXED_CTRnum()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 10; // 0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 7,0) > 1) {
@@ -141,7 +141,7 @@ bool cpuid_perf_global_ctrl_EN_FIXED_CTRnum()
 
 bool cpuid_misc_enable_turboBoost()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 1, 1) == 1) {
@@ -154,7 +154,7 @@ bool cpuid_misc_enable_turboBoost()
 
 bool cpuid_misc_enable_xTPRmessageDisable()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 1;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rcx, 14, 14) == 1) {
@@ -167,7 +167,7 @@ bool cpuid_misc_enable_xTPRmessageDisable()
 
 bool cpuid_misc_enable_XDbitDisable()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 2147483649; // 80000001H
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rdx, 20, 20) == 1) {
@@ -182,7 +182,7 @@ bool cpuid_misc_enable_XDbitDisable()
 
 bool cpuid_clock_mod_extended()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 5,5) == 1) {
@@ -197,7 +197,7 @@ bool cpuid_clock_mod_extended()
 
 bool cpuid_therm_stat_therm_thresh()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 1;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rcx, 8, 8) == 1) {
@@ -210,7 +210,7 @@ bool cpuid_therm_stat_therm_thresh()
 
 bool cpuid_therm_stat_powerlimit()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 4, 4) == 1) {
@@ -223,7 +223,7 @@ bool cpuid_therm_stat_powerlimit()
 
 bool cpuid_therm_stat_readout()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 0, 0) == 1) {
@@ -236,7 +236,7 @@ bool cpuid_therm_stat_readout()
 
 bool cpuid_therm_interrupt_powerlimit()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 4, 4) == 1) {
@@ -249,7 +249,7 @@ bool cpuid_therm_interrupt_powerlimit()
 
 bool cpuid_pkg_therm_Stat_AND_Interrupt()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 6;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	if(MASK_VAL(rax, 6, 6) == 1) {
@@ -263,9 +263,9 @@ bool cpuid_pkg_therm_Stat_AND_Interrupt()
 ///////////////////////////////////////////////////////////////////////////////////////////////////////
 //--------------------------------------General Machine Info-------------------------------------------
 
-uint32_t cpuid_maxleaf()
+uint64_t cpuid_maxleaf()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf=0;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return rax;
@@ -273,7 +273,7 @@ uint32_t cpuid_maxleaf()
 
 void cpuid_printVendorID()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf=0, i=0;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	for(i = 0; i < 32; i+=8)
@@ -296,7 +296,7 @@ void cpuid_printVendorID()
 
 int cpuid_pkg_maxPhysicalProcessorCores()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leafa = 4, leafc = 0;
 	cpuidInput_rax_rcx(leafa, leafc, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rax, 31,26) + 1;
@@ -304,7 +304,7 @@ int cpuid_pkg_maxPhysicalProcessorCores()
 
 int cpuid_pkg_maxLogicalProcessors()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf = 1;
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rax, 23,16);
@@ -313,7 +313,7 @@ int cpuid_pkg_maxLogicalProcessors()
 // The 2 functions below are for the fixed performance counters. See Manual Vol. 18.2.2.1 
 int cpuid_num_fixed_perf_counters()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf= 10; //0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rdx,4,0);
@@ -321,7 +321,7 @@ int cpuid_num_fixed_perf_counters()
 
 int cpuid_width_fixed_perf_counters()
 {
-	uint32_t rax, rbx, rcx, rdx;
+	uint64_t rax, rbx, rcx, rdx;
 	int leaf= 10; //0A
 	cpuid(leaf, &rax, &rbx, &rcx, &rdx);
 	return MASK_VAL(rdx,12,5);
